Restructure data unit tests to follow the new data structure
