+++ tests/unit/data/datamodule/base/image.py
@@ -7,8 +7,7 @@ import numpy as np
 import pytest
 
 from anomalib.data import AnomalibDataModule
-
-from .base import _TestAnomalibDataModule
+from tests.unit.data.datamodule.base.base import _TestAnomalibDataModule
 
 
 class _TestAnomalibImageDatamodule(_TestAnomalibDataModule):
