+++ tests/unit/data/datamodule/base/video.py
@@ -9,8 +9,7 @@ import pytest
 import torch
 
 from anomalib.data import AnomalibDataModule
-
-from .base import _TestAnomalibDataModule
+from tests.unit.data.datamodule.base.base import _TestAnomalibDataModule
 
 
 class _TestAnomalibVideoDatamodule(_TestAnomalibDataModule):
