+++ tests/unit/data/datamodule/image/test_kolektor.py
@@ -1,4 +1,4 @@
-"""Unit Tests - MVTec3D Datamodule."""
+"""Unit Tests - Kolektor Datamodule."""
 
 # Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
@@ -8,26 +8,25 @@ from pathlib import Path
 import pytest
 
 from anomalib import TaskType
-from anomalib.data import MVTec3D
-from tests.unit.data.base import _TestAnomalibDepthDatamodule
+from anomalib.data import Kolektor
+from tests.unit.data.datamodule.base.image import _TestAnomalibImageDatamodule
 
 
-class TestMVTec3D(_TestAnomalibDepthDatamodule):
-    """MVTec Datamodule Unit Tests."""
+class TestKolektor(_TestAnomalibImageDatamodule):
+    """Kolektor Datamodule Unit Tests."""
 
     @pytest.fixture()
     @staticmethod
-    def datamodule(dataset_path: Path, task_type: TaskType) -> MVTec3D:
-        """Create and return a Folder 3D datamodule."""
-        _datamodule = MVTec3D(
-            root=dataset_path / "mvtec_3d",
-            category="dummy",
+    def datamodule(dataset_path: Path, task_type: TaskType) -> Kolektor:
+        """Create and return a BTech datamodule."""
+        _datamodule = Kolektor(
+            root=dataset_path / "kolektor",
             task=task_type,
             image_size=256,
             train_batch_size=4,
             eval_batch_size=4,
-            num_workers=0,
         )
+
         _datamodule.prepare_data()
         _datamodule.setup()
 
@@ -37,4 +36,4 @@ class TestMVTec3D(_TestAnomalibDepthDatamodule):
     @staticmethod
     def fxt_data_config_path() -> str:
         """Return the path to the test data config."""
-        return "configs/data/mvtec_3d.yaml"
+        return "configs/data/kolektor.yaml"
