+++ tests/unit/data/datamodule/image/test_mvtec.py
@@ -1,4 +1,4 @@
-"""Unit Tests - Visa Datamodule."""
+"""Unit Tests - MVTec Datamodule."""
 
 # Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
@@ -8,25 +8,23 @@ from pathlib import Path
 import pytest
 
 from anomalib import TaskType
-from anomalib.data import Visa
-from tests.unit.data.base.image import _TestAnomalibImageDatamodule
+from anomalib.data import MVTec
+from tests.unit.data.datamodule.base.image import _TestAnomalibImageDatamodule
 
 
-class TestVisa(_TestAnomalibImageDatamodule):
-    """Visa Datamodule Unit Tests."""
+class TestMVTec(_TestAnomalibImageDatamodule):
+    """MVTec Datamodule Unit Tests."""
 
     @pytest.fixture()
     @staticmethod
-    def datamodule(dataset_path: Path, task_type: TaskType) -> Visa:
-        """Create and return a Avenue datamodule."""
-        _datamodule = Visa(
-            root=dataset_path,
+    def datamodule(dataset_path: Path, task_type: TaskType) -> MVTec:
+        """Create and return a MVTec datamodule."""
+        _datamodule = MVTec(
+            root=dataset_path / "mvtec",
             category="dummy",
-            image_size=256,
+            task=task_type,
             train_batch_size=4,
             eval_batch_size=4,
-            num_workers=0,
-            task=task_type,
         )
         _datamodule.prepare_data()
         _datamodule.setup()
@@ -37,4 +35,4 @@ class TestVisa(_TestAnomalibImageDatamodule):
     @staticmethod
     def fxt_data_config_path() -> str:
         """Return the path to the test data config."""
-        return "configs/data/visa.yaml"
+        return "configs/data/mvtec.yaml"
