+++ tests/unit/data/datamodule/image/test_visa.py
@@ -1,4 +1,4 @@
-"""Unit Tests - BTech Datamodule."""
+"""Unit Tests - Visa Datamodule."""
 
 # Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
@@ -8,26 +8,26 @@ from pathlib import Path
 import pytest
 
 from anomalib import TaskType
-from anomalib.data import BTech
-from tests.unit.data.base.image import _TestAnomalibImageDatamodule
+from anomalib.data import Visa
+from tests.unit.data.datamodule.base.image import _TestAnomalibImageDatamodule
 
 
-class TestBTech(_TestAnomalibImageDatamodule):
-    """MVTec Datamodule Unit Tests."""
+class TestVisa(_TestAnomalibImageDatamodule):
+    """Visa Datamodule Unit Tests."""
 
     @pytest.fixture()
     @staticmethod
-    def datamodule(dataset_path: Path, task_type: TaskType) -> BTech:
-        """Create and return a BTech datamodule."""
-        _datamodule = BTech(
-            root=dataset_path / "btech",
+    def datamodule(dataset_path: Path, task_type: TaskType) -> Visa:
+        """Create and return a Avenue datamodule."""
+        _datamodule = Visa(
+            root=dataset_path,
             category="dummy",
-            task=task_type,
             image_size=256,
             train_batch_size=4,
             eval_batch_size=4,
+            num_workers=0,
+            task=task_type,
         )
-
         _datamodule.prepare_data()
         _datamodule.setup()
 
@@ -37,4 +37,4 @@ class TestBTech(_TestAnomalibImageDatamodule):
     @staticmethod
     def fxt_data_config_path() -> str:
         """Return the path to the test data config."""
-        return "configs/data/btech.yaml"
+        return "configs/data/visa.yaml"
