+++ tests/unit/data/datamodule/video/test_avenue.py
@@ -1,4 +1,4 @@
-"""Unit Tests - UCSDped Datamodule."""
+"""Unit Tests - Avenue Datamodule."""
 
 # Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
@@ -8,12 +8,12 @@ from pathlib import Path
 import pytest
 
 from anomalib import TaskType
-from anomalib.data import UCSDped
-from tests.unit.data.base.video import _TestAnomalibVideoDatamodule
+from anomalib.data import Avenue
+from tests.unit.data.datamodule.base.video import _TestAnomalibVideoDatamodule
 
 
-class TestUCSDped(_TestAnomalibVideoDatamodule):
-    """UCSDped Datamodule Unit Tests."""
+class TestAvenue(_TestAnomalibVideoDatamodule):
+    """Avenue Datamodule Unit Tests."""
 
     @pytest.fixture()
     @staticmethod
@@ -23,18 +23,19 @@ class TestUCSDped(_TestAnomalibVideoDatamodule):
 
     @pytest.fixture()
     @staticmethod
-    def datamodule(dataset_path: Path, task_type: TaskType, clip_length_in_frames: int) -> UCSDped:
-        """Create and return a UCSDped datamodule."""
-        _datamodule = UCSDped(
-            root=dataset_path / "ucsdped",
-            category="dummy",
+    def datamodule(dataset_path: Path, task_type: TaskType, clip_length_in_frames: int) -> Avenue:
+        """Create and return a Avenue datamodule."""
+        _datamodule = Avenue(
+            root=dataset_path / "avenue",
+            gt_dir=dataset_path / "avenue" / "ground_truth_demo",
             clip_length_in_frames=clip_length_in_frames,
-            task=task_type,
             image_size=256,
+            task=task_type,
+            num_workers=0,
             train_batch_size=4,
             eval_batch_size=4,
-            num_workers=0,
         )
+
         _datamodule.prepare_data()
         _datamodule.setup()
 
@@ -44,4 +45,4 @@ class TestUCSDped(_TestAnomalibVideoDatamodule):
     @staticmethod
     def fxt_data_config_path() -> str:
         """Return the path to the test data config."""
-        return "configs/data/ucsd_ped.yaml"
+        return "configs/data/avenue.yaml"
