+++ tests/unit/data/datamodule/video/test_shanghaitech.py
@@ -1,4 +1,4 @@
-"""Unit Tests - Avenue Datamodule."""
+"""Unit Tests - ShanghaiTech Datamodule."""
 
 # Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
@@ -8,12 +8,12 @@ from pathlib import Path
 import pytest
 
 from anomalib import TaskType
-from anomalib.data import Avenue
-from tests.unit.data.base.video import _TestAnomalibVideoDatamodule
+from anomalib.data import ShanghaiTech
+from tests.unit.data.datamodule.base.video import _TestAnomalibVideoDatamodule
 
 
-class TestAvenue(_TestAnomalibVideoDatamodule):
-    """Avenue Datamodule Unit Tests."""
+class TestShanghaiTech(_TestAnomalibVideoDatamodule):
+    """ShanghaiTech Datamodule Unit Tests."""
 
     @pytest.fixture()
     @staticmethod
@@ -23,17 +23,17 @@ class TestAvenue(_TestAnomalibVideoDatamodule):
 
     @pytest.fixture()
     @staticmethod
-    def datamodule(dataset_path: Path, task_type: TaskType, clip_length_in_frames: int) -> Avenue:
-        """Create and return a Avenue datamodule."""
-        _datamodule = Avenue(
-            root=dataset_path / "avenue",
-            gt_dir=dataset_path / "avenue" / "ground_truth_demo",
+    def datamodule(dataset_path: Path, task_type: TaskType, clip_length_in_frames: int) -> ShanghaiTech:
+        """Create and return a Shanghai datamodule."""
+        _datamodule = ShanghaiTech(
+            root=dataset_path / "shanghaitech",
+            scene=1,
             clip_length_in_frames=clip_length_in_frames,
-            image_size=256,
-            task=task_type,
-            num_workers=0,
+            image_size=(256, 256),
             train_batch_size=4,
             eval_batch_size=4,
+            num_workers=0,
+            task=task_type,
         )
 
         _datamodule.prepare_data()
@@ -45,4 +45,4 @@ class TestAvenue(_TestAnomalibVideoDatamodule):
     @staticmethod
     def fxt_data_config_path() -> str:
         """Return the path to the test data config."""
-        return "configs/data/avenue.yaml"
+        return "configs/data/shanghaitech.yaml"
