+++ tests/unit/data/datamodule/video/test_ucsdped.py
@@ -1,4 +1,4 @@
-"""Unit Tests - ShanghaiTech Datamodule."""
+"""Unit Tests - UCSDped Datamodule."""
 
 # Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
@@ -8,12 +8,12 @@ from pathlib import Path
 import pytest
 
 from anomalib import TaskType
-from anomalib.data import ShanghaiTech
-from tests.unit.data.base.video import _TestAnomalibVideoDatamodule
+from anomalib.data import UCSDped
+from tests.unit.data.datamodule.base.video import _TestAnomalibVideoDatamodule
 
 
-class TestShanghaiTech(_TestAnomalibVideoDatamodule):
-    """ShanghaiTech Datamodule Unit Tests."""
+class TestUCSDped(_TestAnomalibVideoDatamodule):
+    """UCSDped Datamodule Unit Tests."""
 
     @pytest.fixture()
     @staticmethod
@@ -23,19 +23,18 @@ class TestShanghaiTech(_TestAnomalibVideoDatamodule):
 
     @pytest.fixture()
     @staticmethod
-    def datamodule(dataset_path: Path, task_type: TaskType, clip_length_in_frames: int) -> ShanghaiTech:
-        """Create and return a Shanghai datamodule."""
-        _datamodule = ShanghaiTech(
-            root=dataset_path / "shanghaitech",
-            scene=1,
+    def datamodule(dataset_path: Path, task_type: TaskType, clip_length_in_frames: int) -> UCSDped:
+        """Create and return a UCSDped datamodule."""
+        _datamodule = UCSDped(
+            root=dataset_path / "ucsdped",
+            category="dummy",
             clip_length_in_frames=clip_length_in_frames,
-            image_size=(256, 256),
+            task=task_type,
+            image_size=256,
             train_batch_size=4,
             eval_batch_size=4,
             num_workers=0,
-            task=task_type,
         )
-
         _datamodule.prepare_data()
         _datamodule.setup()
 
@@ -45,4 +44,4 @@ class TestShanghaiTech(_TestAnomalibVideoDatamodule):
     @staticmethod
     def fxt_data_config_path() -> str:
         """Return the path to the test data config."""
-        return "configs/data/shanghaitech.yaml"
+        return "configs/data/ucsd_ped.yaml"
