--- tests/unit/data/image/__init__.py
@@ -1,4 +1,0 @@
-"""Unit tests - Image Datamodules."""
-
-# Copyright (C) 2023-2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
