--- tests/unit/data/video/__init__.py
@@ -1,4 +1,0 @@
-"""Unit tests - Video Datamodules."""
-
-# Copyright (C) 2023-2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
